pyproject.toml
[project]
name = "traderminion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django==5.2.8",
    "django-cors-headers==4.3.1",
    "djangorestframework==3.14.0",
    "pillow>=11.0.0",
]
